added conditional loss and two channels

+++ src/config.py
@@ -6,6 +6,8 @@ from pathlib import Path, WindowsPath
 from shutil import rmtree
 from termcolor import colored
 
+import numpy as np
+
 # from synth_config import BASIC_FLOW
 from typing import Dict, List
 
@@ -75,7 +77,7 @@ class Config:
     freq_reinforce_loss_factor: float = 1e5
 
     multi_spectral_loss_spec_type: str = 'BOTH'
-    multi_spectral_loss_preset: str = 'cumsum_time'
+    multi_spectral_loss_preset: str = 'cumsum_time_high_fft'
 
     # Debug
     debug_mode: bool = False
@@ -131,8 +133,8 @@ class Config:
 
 @dataclass
 class DatasetConfig:
-    dataset_size: int = 100
-    batch_size: int = 100
+    dataset_size: int = 50000
+    batch_size: int = 1000
     num_epochs_to_print_stats: int = 100
     train_parameters_file: str = None
     train_audio_dir: str = None
@@ -159,12 +161,12 @@ class DatasetConfig:
 
 @dataclass
 class ModelConfig:
-    preset: str = 'OSC_AMPLITUDE_SHAPER'
+    preset: str = 'NON_SINE_LFO'
     model_type: str = 'simple'
     backbone: str = 'resnet'
     batch_size: int = 128
     num_epochs: int = 20
-    learning_rate: float = 3e-4
+    learning_rate: float = 3e-5
     optimizer_weight_decay: float = 0
     optimizer_scheduler_lr: float = 0
     optimizer_scheduler_gamma: float = 0.1
@@ -174,7 +176,7 @@ class ModelConfig:
 
 @dataclass
 class SynthConfig:
-    preset: str = 'OSC_AMPLITUDE_SHAPER'
+    preset: str = 'BASIC_FLOW_NO_ADSR_NO_FILTER'
     wave_type_dict = {"sine": 0,
                       "square": 1,
                       "sawtooth": 2}
@@ -202,8 +204,9 @@ class SynthConfig:
     # Modular synth possible modules from synth_modules.py
     modular_synth_operations = ['osc', 'fm', 'lfo', 'mix', 'filter', 'env_adsr']
     modular_synth_params = {'osc': ['amp', 'freq', 'waveform'],
-                            'lfo': ['freq', 'waveform'],
-                            'fm': ['amp_c', 'freq_c', 'waveform', 'mod_index'],
+                            'lfo_sine': ['freq'],
+                            'lfo_non_sine': ['freq', 'waveform'],
+                            'fm': ['freq_c', 'waveform', 'mod_index'],
                             'mix': None,
                             'filter': ['filter_freq', 'filter_type'],
                             'env_adsr': ['attack_t', 'decay_t', 'sustain_t', 'sustain_level', 'release_t'],
@@ -220,21 +223,28 @@ class SynthConfig:
         self.osc_freq_dic_inv = {v: k for k, v in self.osc_freq_dic.items()}
         self.oscillator_freq = self.osc_freq_list[-1] + self.margin
 
+        self.all_params_presets = {
+            'lfo': {'freq': np.asarray([0.5] + [k+1 for k in range(self.max_lfo_freq)])},
+            'fm': {'freq_c': np.asarray(self.osc_freq_list),
+                   'mod_index': np.linspace(0, self.max_mod_index, 16)},
+            'filter': {'filter_freq': np.asarray([100*1.4**k for k in range(14)])}
+        }
+
 
 def configure_experiment(exp_name: str, dataset_name: str):
 
     project_root = os.path.join(EXP_ROOT, 'current', exp_name, '')
 
-    # if os.path.isdir(project_root):
-    #     overwrite = input(colored(f"Folder {project_root} already exists. Overwrite previous experiment (Y/N)?"
-    #                               f"\n\tThis will delete all files related to the previous run!",
-    #                               'yellow'))
-    #     if overwrite.lower() != 'y':
-    #         print('Exiting...')
-    #         exit()
+    if os.path.isdir(project_root):
+        overwrite = input(colored(f"Folder {project_root} already exists. Overwrite previous experiment (Y/N)?"
+                                  f"\n\tThis will delete all files related to the previous run!",
+                                  'yellow'))
+        if overwrite.lower() != 'y':
+            print('Exiting...')
+            exit()
 
-    print("Deleting previous experiment...")
-    rmtree(project_root)
+        print("Deleting previous experiment...")
+        rmtree(project_root)
 
     cfg = Config(project_root)
     synth_cfg = SynthConfig()
